MueLu: split HierarchyManager into decl/def file
